Windows: Fix, icon group entries were wrongly sized
* The id field would also potentially be wrongly set.

+++ nuitka/PostProcessing.py
@@ -82,7 +82,7 @@ class IconGroupDirectoryEntry(ctypes.Structure):
         ("planes", ctypes.c_short),
         ("bit_count", ctypes.c_short),
         ("image_size", ctypes.c_int),
-        ("id", ctypes.c_int),
+        ("id", ctypes.c_short),
     )
